Symbolic (#22)
* Merging into symbolic (#21)

* Installing docsify

* Creating website structure

* getting ready for writing documentation

* Fixing broken link to nodejs website

* Adding configurations and first lines of doc

* Adding typescript syntax highlighting

* Adding index and sidebar for v0.2.1

* Adding more navigation links

* 0.2.3

* Adding documentation structure for definitions

* Modifying main sidebar

* Doc (#19)

* Adding field descriptions

* Adding quantity field for evaluable

* Adding the Variable interface

* Adding docs for Constant interface

* Adding docs for Expression interface

* Adding docs for isEvaluable

* Adding docs for isVariable, isConstant, isExpression

* Fixing typo

* Adding initial docs for Scalar

* Completing docs for Scalar

* Changing page layout

* Adding docsfor Constant interface

* docs: Expression and Variable interfaces

* docs: Evaluable and Token interfaces

* docs: abstract class Scalar

* docs: class Scalar.Constant

* docs: Fixing interface dependency

* Cleaning up dependency structure

* docs: Further cleaning up dependencies

* Fixing minor mistake

* docs: Adding class Scalar.Variable

* docs: Adding Scalar.Expression

* docs: adding sidebar

* Adding extra structure to page layout

* Adding styling

* Moving style to separate file

* Minor layout changes

* Migrating from docsify to typedoc

* Updating README

* Updating docs

* 0.2.4

* ignoring docs folder

* 0.2.5

* Adding a zero element for scalars

* Adding new constants and errors

* Adding interval object

* Removing unnecessary bits and restructuring directory
