+++ src/index.ts
@@ -1,7 +1,10 @@
-export { Matrix } from "./matrix";
-export { Vector, __ } from "./vector";
-export { Scalar } from "./scalar";
-export { Evaluable, isEvaluable, Constant, isConstant, Variable, isVariable, Expression, isExpression, Operator } from "./core/definitions";
-export { ExpressionBuilder } from "./core/expression";
-export { BinaryOperator, isBinaryOperator } from "./core/operators/binary";
-export { neg, sin, cos, tan, asin, acos, atan, sinh, cosh, tanh, asinh, acosh, atanh, log, ln, exp, abs, floor, ceil, UnaryOperator, isUnaryOperator } from "./core/operators/unary";
+export * from "./matrix";
+export * from "./vector";
+export * from "./scalar";
+export * from "./core/definitions";
+export * from "./core/expression";
+export * from "./core/operators/binary";
+export * from "./core/operators/unary";
+export * from "./core/math/functions";
+export * from "./core/errors";
+export * from "./core/interval";
